+++ generate_version_info.py
@@ -2,7 +2,7 @@ import requests
 import re
 
 def get_version():
-    return "1.0.7.9"  # Версия по умолчанию
+    return "1.0.8.0"  # Версия по умолчанию
 
 def generate_version_info():
     version = get_version()
@@ -29,7 +29,7 @@ VSVersionInfo(
         StringStruct(u'FileDescription', u'IB Launcher'),
         StringStruct(u'FileVersion', u'{version}'),
         StringStruct(u'InternalName', u'IB-Launcher'),
-        StringStruct(u'LegalCopyright', u'Copyright (c) 2024 Igrobar'),
+        StringStruct(u'LegalCopyright', u'Copyright (c) 2024-2025 Igrobar'),
         StringStruct(u'OriginalFilename', u'IB-Launcher.exe'),
         StringStruct(u'ProductName', u'IB Launcher'),
         StringStruct(u'ProductVersion', u'{version}')])